v2: Debounce unnecessary Service writes to Consul
Endpoints Controller v2 should not write services every time endpoints
change, as this can occur far more frequently and be subject to rapid
updates in the case of large clusters or flapping pod health.

Instead, attempt to read an existing service resource, and write only if
the resource would be updated by the controller, or the read fails
(optimistically write as a best-effort attempt rather than blocking
indefinitely on reads).