Add xds-fetch-timeout-ms annotation handling.
This new annotation allows users to specify the xds_fetch_timeout_ms
configuration during service registration. Use of this annotation
is not encouraged in most situations, but it is necessary for certain
circumstances where a large number of upstreams are configured on a
single Envoy proxy.

Without this setting, anytime it took longer than 15s (the default)
to configure Envoy's ClusterLoadAssignments, there was potential for
endpoints to timeout during a fetch. This would result in strange
missing endpoints whenever the xDS connection is re-established after
some future disruption.